requirements.txt
numpy
pandas
matplotlib
scipy
EMD-signal
python-dotenv
requests
typer
pytest
pyarrow
plotly
scikit-learn
dash
markdown 
weasyprint
pillow
fpdf